+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - cli: `todo delay` and `todo reschedule` now accept multiple ids and `--time` change to an argument
+- cli: `event new`, `event edit`, `todo new`, and `todo edit` now invoke the TUI if any required
+  fields are missing
 - core: Add static lifecycle for all aim methods
 
 ### Fixed